making timezones handle :30 and :45 offsets correctly

+++ src/HydroDatas.h
@@ -36,7 +36,7 @@ struct HydroSystemData : public HydroData {
     Hydro_DisplayOutputMode dispOutMode;                    // System display output mode
     Hydro_ControlInputMode ctrlInMode;                      // System control input mode 
     char systemName[HYDRO_NAME_MAXSIZE];                    // System name
-    int8_t timeZoneOffset;                                  // Timezone offset
+    int16_t timeZoneOffset;                                 // Timezone offset *100 (accounts for :30 and :45)
     uint16_t pollingInterval;                               // Sensor polling interval, in milliseconds
     Hydro_Autosave autosaveEnabled;                         // Autosave enabled
     Hydro_Autosave autosaveFallback;                        // Autosave fallback
